restructure polynomial crate to include eval form multilinear poly
