+++ gkr/src/gate.rs
@@ -1,4 +1,4 @@
-use polynomial::multilinear_poly::binary_string;
+use polynomial::multilinear::coefficient_form::binary_string;
 
 /// Represents a gate wiring (2 inputs 1 output)
 #[derive(Clone, Debug, PartialEq)]
