+++ polynomial/src/lib.rs
@@ -2,10 +2,7 @@ use ark_ff::PrimeField;
 
 use self::univariate_poly::UnivariatePolynomial;
 
-// pub mod multilinear_extension;
-pub mod multilinear_poly;
-mod multilinear_poly_2;
-mod pairing_index;
+pub mod multilinear;
 pub mod univariate_poly;
 
 pub trait Polynomial<F: PrimeField>: Clone {
