+++ polynomial/src/multilinear/coefficient_form.rs
@@ -510,7 +510,7 @@ pub fn bit_count_for_n_elem(size: usize) -> usize {
 
 #[cfg(test)]
 mod tests {
-    use crate::multilinear_poly::{
+    use crate::multilinear::coefficient_form::{
         mapping_instruction_from_variable_presence, selector_to_index, to_power_of_two,
         MultiLinearPolynomial,
     };
